RY_Study/DeeplApi/backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
pandas
requests
django-allauth
